change put$delete = post

+++ app/Http/Controllers/DriverController.php
@@ -15,7 +15,8 @@ use Symfony\Component\HttpKernel\Exception\HttpException;
 class DriverController extends Controller
 {
     public function showAllDriversData()
-    {$bigArray = array();
+    {
+        $bigArray = array();
         $tempArray = array();
         $reviews = array();
         $routeGroup = array();
@@ -89,14 +90,14 @@ class DriverController extends Controller
 
             file_put_contents($path, $decoded);
         }
-        $driver = $request->isMethod('put') ? Driver::findOrFail($request->id)
-        : new Driver;
 
-        $path_old_photo = '';
-
-        if ($request->isMethod('put')) {
-            $path_old_photo = $driver->photo;
+        if (isset($request->id)) {
+            $driver = Driver::findOrFail($request->id);
+        } else {
+            $driver = new Driver;
         }
+
+        $path_old_photo = $driver->photo;
 
         $driver->id = $request->input('id');
         $driver->name = $request->input('name');
@@ -110,7 +111,9 @@ class DriverController extends Controller
         $car = $driver->car()->create();
         $car->carphoto()->create();
 
-        if ($path_old_photo) {
+        \Log::info($path_old_photo);
+
+        if ($path_old_photo && strpos($path_old_photo, "https://") !== 0) {
             unlink(public_path() . $path_old_photo);
         }
